+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPN6KHD5XJ6TVA0NPTTQF4B"
+        "01KNPWKHJVNGWH8Y65XDVGJSKR"
     )
     r"""The unique identifier of the data source"""