+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPN6KHFEX6GP7E6FA1D4GY6"
+        "01KNPWKHJY93Y4YJMSBJ550C92"
     )
     r"""The unique identifier of the data source"""