+++ packages/orq-rc/RELEASES.md
@@ -6959,3 +6959,13 @@ Based on:
 - [python v4.7.0-rc.56] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.56] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.56 - packages/orq-rc
+
+## 2026-04-08 15:53:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.57] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.57] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.57 - packages/orq-rc